requirements.txt
streamlit
pandas
numpy
openai
python-dotenv
scikit-learn

# pip install -r requirements.txt
